+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [2.09] - 2025-12-15
+### Added
+- If PCToBCExchangeRate is zero, set it to 1.0 to fix proper currency conversion rate
+- Error/Waring reporting
+
+### Fixed
+- Trace log for import filesize data for Application Insights
+
+
 ## [2.06] - 2025-05-16
 ### Added
 - Add function to clean CVS files with wrong JSON elements included in the file
@@ -8,10 +17,12 @@
 ### Changed
 - Refactor code with best practices
 
+
 ## [2.03] - 2024-02-27
 ### Changed
 - Change selector for subscription - in billing there is new field named Entitlement, and old SubscriptionId is for Azure plan ID
 - Change view for better visibility of data
+
 
 ## [2.02] - 2024-02-15
 ### Fixed
@@ -20,9 +31,11 @@
 ### Changed
 - Update the version of bootstrap to version 5.3.2
 
+
 ## [2.01] - 2023-04-27
 ### Added
 - Add Azure Application Insights, report each subscription separately
+
 
 ## [2.00] - 2023-04-27### Added
 ### Added
@@ -30,15 +43,16 @@
 ### Removed
 - Add HTTP server using Flesk
 
+
 ## [1.02] - 2023-04-12
 ### Added
 - HTML output and add parser for input arguments
+
 
 ## [1.01] - 2022-11-21
 ### Added
 - CSV Support
 
+
 ## [1.00] - 2022-07-15
 ### Initial public version
-
-